Fix e2e scripts
